Add files via upload
Subfolder "Check mathematical expressions". These are codes to numerically verify mathematical expressions that I derived during the development of the numerical method. Many are probably no longer of interest, but I upload them all for reference. A few, such as proofs for Chebyshev polynomials, are referred to in the thesis. Again, references to custom header files may need updating.